% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewparams.R
\name{build_viewparams}
\alias{build_viewparams}
\title{String that can be passed to viewParams}
\usage{
build_viewparams(
  mrgid = NULL,
  geometry = NULL,
  dasid = NULL,
  startdate = NULL,
  enddate = NULL,
  aphiaid = NULL,
  functional_groups = NULL,
  cites = NULL,
  habitats_directive = NULL,
  iucn_red_list = NULL,
  msdf_indicators = NULL
)
}
\arguments{
\item{mrgid}{Marine Regions Gazetteer unique identifier}

\item{dasid}{IMIS dataset unique identifier}

\item{startdate}{Start date of occurrences}

\item{enddate}{End date of occurrences}

\item{aphiaid}{WoRMS taxon unique identifier}

\item{polygon}{Polygon of the region of interest as WKT or sf object. Draw with eurobis_draw_polygon.}
}
\value{
a string that can be passed to viewParams in a WFS request to EMODnet-Biology
}
\description{
String that can be passed to viewParams
}
\examples{
build_viewparams()
build_viewparams(mrgid = 8364)
build_viewparams(polygon = 'POLYGON((-2 52,-2 58,9 58,9 52,-2 52))')
build_viewparams(mrgid = 8364, polygon = 'POLYGON((-2 52,-2 58,9 58,9 52,-2 52))')
build_viewparams(dasid = 216)
build_viewparams(dasid = 216, startdate = "2000-01-01", enddate = "2022-01-31")
build_viewparams(mrgid = 8364, polygon = 'POLYGON((-2 52,-2 58,9 58,9 52,-2 52))', 
                 dasid = 216, startdate = "2000-01-01", enddate = "2022-01-31", aphiaid = c(104108, 148947))
}
