% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.r
\name{cleandataframe}
\alias{cleandataframe}
\title{remove empty collumns and make all collumns into characters}
\usage{
cleandataframe(x)
}
\description{
remove empty collumns and make all collumns into characters
}
