% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createwfsurls.r
\name{createwfsurls}
\alias{createwfsurls}
\title{Helpfunction to create the wfsurl}
\usage{
createwfsurls(geourl = NA, dasid = NA, aphiaid = NA,
  startyear = "1850", endyear = NA, type = "full")
}
\description{
Helpfunction to create the wfsurl
}
