% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{eurobis_download_toolbox}
\alias{eurobis_download_toolbox}
\title{Browse the LifeWatch/EMODnet-Biology Download Toolbox on your browser}
\usage{
eurobis_download_toolbox()
}
\value{
NULL, only opens the Download Toolbox in your browser.
}
\description{
Browse the LifeWatch/EMODnet-Biology Download Toolbox on your browser
}
\examples{
eurobis_download_toolbox()
}
