% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{eurobis_list_datasets}
\alias{eurobis_list_datasets}
\title{List EurOBIS datasets}
\usage{
eurobis_list_datasets()
}
\value{
a tibble with the dataset ID and the dataset title that can be queried with the 
eurobis package. See details section.
}
\description{
List EurOBIS datasets
}
\details{
The datasets of EurOBIS are documented in the Integrated Marine Information System (IMIS)
If you want to get more information about a dataset, you can look up the dataset 
in https://www.vliz.be/en/imis?module=dataset.

Note you can pass the dataset ID to the url. e.g: 
https://www.vliz.be/en/imis?module=dataset&dasid=216 

If you prefer a machine-readable format, add `&show=json`
https://www.vliz.be/en/imis?module=dataset&dasid=216&show=json

You can read this easily in R, e.g:
jsonlite::fromJSON("https://www.vliz.be/en/imis?module=dataset&dasid=216&show=json")
}
\examples{
\dontrun{eurobis_list_datasets()}
}
