% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{eurobis_list_regions}
\alias{eurobis_list_regions}
\title{List Marine Regions}
\usage{
eurobis_list_regions()
}
\value{
a tibble with the Marine Regions names and unique identifier MRGID. See details section.
}
\description{
List Marine Regions
}
\details{
The Marine Regions Gazetteer is the geographical backbone of EurOBIS.

The MRGID is the Marine Regions Gazetteer Unique Identifier. This is a persistent ID that won't be ever deleted. 
See https://marineregions.org/mrgid.php for further information.

You can find more info about each marine region by searching in the Marine Regions website: 
https://marineregions.org/gazetteer.php?p=search

Or you can use the mregion2 R package (on development): 
https://github.com/lifewatch/mregions2

E.g.:
mregions2::mr_gaz_record(25600)
}
\examples{
\dontrun{eurobis_list_regions()}
}
