% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eurobis_map_draw}
\alias{eurobis_map_draw}
\title{Draw interactively a polygon and get it as Well Known Text}
\usage{
eurobis_map_draw()
}
\value{
a string with a polygon as Well Known Text
}
\description{
Draw interactively a polygon and get it as Well Known Text
}
\examples{
\dontrun{wkt <- eurobis_map_draw()}
}
