% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eurobis_map_mr}
\alias{eurobis_map_mr}
\title{Interactive map with Marine Regions products on your viewer. Click to get the MRGID and more info}
\usage{
eurobis_map_mr(layer = "eez")
}
\arguments{
\item{layer}{c('eez', 'iho', 'eez_iho'). See details.}
}
\value{
A leaflet map with Marine Regions product that retrieves info by clicking
}
\description{
Interactive map with Marine Regions products on your viewer. Click to get the MRGID and more info
}
\details{
These are Marine Regions products. Find more information in {https://marineregions.org/}
- eez = Exclusive Economic Zone
- iho = International Hydrographic Office areas
- eez_iho = Intersection of the EEZ and IHO areas
}
\examples{
eurobis_map_mr('eez')
eurobis_map_mr('iho')
eurobis_map_mr('eez_iho')
}
