% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{eurobis_name2id}
\alias{eurobis_name2id}
\title{Get the WoRMS AphiaID given a scientific name}
\usage{
eurobis_name2id(taxa)
}
\arguments{
\item{taxa}{The scientific name of a taxon registered in the World Rercord of Marine Species}
}
\value{
a vector of AphiaID
}
\description{
Get the WoRMS AphiaID given a scientific name
}
\examples{
eurobis_name2id(c("Abra alba", "Engraulis encrasicolus"))
eurobis_name2id(c("foo"))
eurobis_name2id(1)
eurobis_name2id("")
eurobis_name2id(c("Abra alba", "foo", NA))
eurobis_name2id(c("Abra alba", "Engraulis encrasicolus"))
eurobis_name2id(c("abra alba", "foo foo2", "caballo loco"))
}
