% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurobis_occurrences.R
\name{eurobis_type_handler}
\alias{eurobis_type_handler}
\title{#' @rdname eurobis_occurrences
#' @export
eurobis_grid_15m <- function(...){
  eurobis_occurrences(type = "grid15m", ...)
}}
\usage{
eurobis_type_handler(type)
}
\arguments{
\item{type}{Type of data, one of c('basic', 'full', 'full_and_parameters')}
}
\description{
#' @rdname eurobis_occurrences
#' @export
eurobis_grid_30m <- function(...){
  eurobis_occurrences(type = "grid30m", ...)
}
}
\details{
#' @rdname eurobis_occurrences
#' @export
eurobis_grid_1d <- function(...){
  eurobis_occurrences(type = "grid1d", ...)
}

Asserts data type and gets the WFS layer name
}
\examples{
\dontrun{
eurobis_type_handler("basic")
eurobis_type_handler("full")
eurobis_type_handler("full_and_parameters")
}
}
