% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcitations.r
\name{getcitations}
\alias{getcitations}
\title{Get citations and licence from IMIS based on dasids}
\usage{
getcitations(dasid)
}
\arguments{
\item{dasid}{mandatory parameter, the dasid of the dataset}
}
\description{
This function gets the citations and DOIs from IMIS using the dasids
}
\examples{
meta <- getcitations(4662)
meta <- getcitations(dasid = c("1884","618", "5780" ))
}
