% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getemodbiodata.R
\name{getemodbiodata}
\alias{getemodbiodata}
\title{Get data from the EMODnet Biology WFS}
\usage{
getemodbiodata(geourl = NA, dasid = NA, aphiaid = NA,
  startyear = NA, endyear = NA, type = "para")
}
\arguments{
\item{geourl}{optional parameter, in case you use the geoserverURL obtained through the EMODnet Biology toolbox http://www.emodnet-biology.eu/toolbox/en/download/occurrence/explore}

\item{dasid}{optional parameter, in case you want to download a single dataset. Use the dasid obtained through the http://www.emodnet-biology.eu/data-catalog (the id from the url)}

\item{aphiaid}{optional parameter, in case you want to download data from a single taxon. uses the id obtained throug ww.marinespecies.org}

\item{startyear}{optional parameter, the earliest year the collected specimen might have been collected}

\item{endyear}{optional parameter, the latest year the collected specimen might have been collected}

\item{type}{indicates which fields you want returned from the WFS (to be implemented)}
}
\description{
This function allows you to download data using aphiaID, dasid or using an geoserverURL you get from the EMODnet Biology toolbox
}
\examples{
downloaddata <- getemodbiodata(dasid = "4662")
downloaddata <- getemodbiodata(aphiaid = "141433")
downloaddata <- getemodbiodata(dasid = c("1884","618", "5780" ), aphiaid = "2036")
getemodbiodata(geourl = "http://geo.vliz.be/geoserver/wfs/ows?service=WFS&version=1.1.0&request=GetFeature&typeName=Dataportal\%3Aeurobis-obisenv&resultType=results&viewParams=where\%3Adatasetid+IN+\%285885\%29\%3Bcontext\%3A0100&propertyName=datasetid\%2Cdatecollected\%2Cdecimallatitude\%2Cdecimallongitude\%2Ccoordinateuncertaintyinmeters\%2Cscientificname\%2Caphiaid\%2Cscientificnameaccepted\%2Cmodified\%2Cinstitutioncode\%2Ccollectioncode\%2Cyearcollected\%2Cstartyearcollected\%2Cendyearcollected\%2Cmonthcollected\%2Cstartmonthcollected\%2Cendmonthcollected\%2Cdaycollected\%2Cstartdaycollected\%2Cenddaycollected\%2Cseasoncollected\%2Ctimeofday\%2Cstarttimeofday\%2Cendtimeofday\%2Ctimezone\%2Cwaterbody\%2Ccountry\%2Cstateprovince\%2Ccounty\%2Crecordnumber\%2Cfieldnumber\%2Cstartdecimallongitude\%2Cenddecimallongitude\%2Cstartdecimallatitude\%2Cenddecimallatitude\%2Cgeoreferenceprotocol\%2Cminimumdepthinmeters\%2Cmaximumdepthinmeters\%2Coccurrenceid\%2Cscientificnameauthorship\%2Cscientificnameid\%2Ctaxonrank\%2Ckingdom\%2Cphylum\%2Cclass\%2Corder\%2Cfamily\%2Cgenus\%2Csubgenus\%2Cspecificepithet\%2Cinfraspecificepithet\%2Caphiaidaccepted\%2Coccurrenceremarks\%2Cbasisofrecord\%2Ctypestatus\%2Ccatalognumber\%2Creferences\%2Crecordedby\%2Cidentifiedby\%2Cyearidentified\%2Cmonthidentified\%2Cdayidentified\%2Cpreparations\%2Csamplingeffort\%2Csamplingprotocol\%2Cqc\%2Ceventid\%2Cparameter\%2Cparameter_value\%2Cparameter_group_id\%2Cparameter_measurementtypeid\%2Cparameter_bodcterm\%2Cparameter_bodcterm_definition\%2Cparameter_standardunit\%2Cparameter_standardunitid\%2Cparameter_imisdasid\%2Cparameter_ipturl\%2Cparameter_original_measurement_type\%2Cparameter_original_measurement_unit\%2Cparameter_conversion_factor_to_standard_unit\%2Cevent\%2Cevent_type\%2Cevent_type_id&outputFormat=csv")
}
