% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEurobisGrid.R
\name{getEurobisGrid}
\alias{getEurobisGrid}
\title{Get data from EurOBIS as Gridded polygon values}
\usage{
getEurobisGrid(aphiaid = 126436, gridsize = "30m")
}
\arguments{
\item{aphiaid}{The taxon id obtained throug ww.marinespecies.org}

\item{gridsize}{The grid size of which you want to get you grid. Possible grid sizes are '6m', '15m', '30m', '1d' (6 minutes, 15 minutes, 30 minutes, 1 degree).}
}
\description{
This function allows you to download a summary geospatial grid, summarizing the number of eurobis occurrences using a species (aphiaID) and grid size
}
\examples{
getEurobisGrid(aphiaid = "141433")
test <- getEurobisGrid(gridsize = '15m')
mapview(test,
        zcol = 'RecordCount',
        lwd = 0)
}
