% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEurobisPoints.R
\name{getEurobisPoints}
\alias{getEurobisPoints}
\title{Get data from EurOBIS as points}
\usage{
getEurobisPoints(aphiaid = 126436)
}
\arguments{
\item{aphiaid}{The taxon id obtained throug www.marinespecies.org (or the r package 'worrms')}
}
\description{
This function allows you to download a summary geospatial grid, summarizing the number of eurobis occurrences using a species (aphiaID) and grid size
}
\examples{
getEurobisPoints(aphiaid = "141433")
test_points <- getEurobisPoints()
mapview(test_points, cex = 1)
}
