% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getoriginaldata.r
\name{getoriginaldata}
\alias{getoriginaldata}
\title{Get original records from IPT based on EMODnet biology output   -  in development}
\usage{
getoriginaldata(ipturls, eurobisdata = NA)
}
\arguments{
\item{ipturls}{mandatory parameter, a vector containing one or more IPTURLS}
}
\description{
This function allows you to download the original data as provided from IPT using the EMODnet Biology download output
}
\examples{
getoriginaldata(ipturls = meta$datasetIPTurl)
}
