% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listemodnetdatasets.R
\name{listemodnetdatasets}
\alias{listemodnetdatasets}
\title{REMODnet BIO datasets}
\usage{
listemodnetdatasets()
}
\description{
This function lists all datasets you can download using REMODnet BIO package
}
\examples{
listemodnetdatasets()
}
