% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listemodnetdatasets.R
\name{listeurobisdatasets}
\alias{listeurobisdatasets}
\title{Eurobis datasets}
\usage{
listeurobisdatasets()
}
\description{
This function lists all datasets you can download using the eurobis package
}
\examples{
listeurobisdatasets()
}
