% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_traits}
\alias{species_traits}
\title{Species traits and classifications important to society}
\format{
A data frame with 32 rows and 4 variables:
\describe{
  \item{category}{traits or important to society}
  \item{group}{subclategory}
  \item{selection}{value of functional group or classification important to society}
  \item{selectid}{ignore - used internally to retrieve data from EurOBIS using WFS}
  ...
}
}
\usage{
species_traits
}
\description{
You can query on the values shown here
}
\keyword{datasets}
