% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurobis_occurrences.R
\name{eurobis_occurrences}
\alias{eurobis_occurrences}
\title{Get EurOBIS data by using the EMODnet-Biology webservices}
\usage{
eurobis_occurrences(
  type,
  url = NULL,
  mrgid = NULL,
  geometry = NULL,
  dasid = NULL,
  startdate = NULL,
  enddate = NULL,
  aphiaid = NULL,
  scientificname = NULL,
  functional_groups = NULL,
  cites = NULL,
  habitats_directive = NULL,
  iucn_red_list = NULL,
  msdf_indicators = NULL,
  paging = FALSE,
  paging_length = 50000,
  ...
)
}
\arguments{
\item{type}{Type of data, one of c('basic', 'full', 'full_and_parameters'). 
More info at: https://www.emodnet-biology.eu/emodnet-data-format}

\item{url}{A WFS request copied from the LifeWatch/EMODnet-Biology Download Toolbox. Use eurobis_download_toolbox()}

\item{mrgid}{Marine Regions Gazetteer unique identifier}

\item{geometry}{a WKT geometry string or sf object with the region of interest.}

\item{dasid}{IMIS dataset unique identifier.}

\item{startdate}{Start date of occurrences.}

\item{enddate}{End date of occurrences.}

\item{aphiaid}{WoRMS taxon unique identifier.}

\item{paging}{If TRUE, the data will be requested on chunks. Use for large requests.}

\item{paging_length}{Size of chunks in number of rows. Default 50K.}

\item{...}{Any parameters to pass to ows4r getFeature() (e.g. cql_filter or parallel)}
}
\value{

}
\description{
Get EurOBIS data by using the EMODnet-Biology webservices
}
\examples{
test <- eurobis_occurrences("basic", dasid = 8045, mrgid = c(5688, 5686))
test <- eurobis_occurrences("full", dasid = 8045)
test <- eurobis_occurrences("full_and_parameters", dasid = 8045)
test <- eurobis_occurrences("basic", dasid = 8045, scientificname = "Zostera marina")
test <- eurobis_occurrences("full", dasid = 8045, functional_groups = "angiosperms")
}
